installs ready for merge

+++ server/bookstack/stats/models.py
@@ -5,8 +5,8 @@ from django.contrib.auth.models import User
 # Create your models here.
 class UserStats(models.Model):
     user_id = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
-    pages_per_day = models.IntegerField(null=True)
-    avg_book_time = models.FloatField(null=True)
+    pages_per_day = models.FloatField(null=True)
+    avg_book_time = models.IntegerField(null=True)
     avg_book_length = models.IntegerField(null=True)
     total_books_read = models.IntegerField(null=True)
     genres = models.JSONField(null=True)
